% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_fun.R
\name{cal_mae}
\alias{cal_mae}
\title{compute mean absolute error}
\usage{
cal_mae(xhat, xtrue, round = FALSE)
}
\arguments{
\item{xhat}{imputed value vector}

\item{xtrue}{true value vector}

\item{round}{whether round the values of \code{xhat} to integers}
}
\value{
computed error
}
\description{
compute mean absolute error
}
