% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_fun.R
\name{cal_mae_scaled}
\alias{cal_mae_scaled}
\title{compute the scaled mean squared error(SMAE) scaled by the MAE of column}
\usage{
cal_mae_scaled(xhat, xobs, xtrue, round = FALSE, reduce = TRUE)
}
\arguments{
\item{xhat}{complete imputed data matrix}

\item{xobs}{incomplete observed data matrix}

\item{xtrue}{complete true data matrix}

\item{round}{whether round the values of \code{xhat} to integers}
}
\value{
a vector with SMAE for each column
}
\description{
For each column, compute the MAE of imputed value divided by the MAE of column median imputation
}
