% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_fun.R
\name{cal_misclass}
\alias{cal_misclass}
\title{compute mis-classification rate for categorical data}
\usage{
cal_misclass(xhat, xobs, xtrue)
}
\arguments{
\item{xhat}{complete imputed data matrix}

\item{xobs}{incomplete observed data matrix}

\item{xtrue}{complete true data matrix}
}
\description{
compute mis-classification rate for categorical data
}
