% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_fun.R
\name{cal_rmse}
\alias{cal_rmse}
\title{compute the root mean squared error}
\usage{
cal_rmse(xhat, xobs = NULL, xtrue, relative = TRUE)
}
\arguments{
\item{xhat}{imputed data matrix or vector}

\item{xobs}{incomplete observed data matrix or vector}

\item{xtrue}{complete true data matrix or vector}

\item{relative}{whether divide the computed error by the Frobenius norm of \code{xtrue}}
}
\value{
computed error
}
\description{
compute the root mean squared error on entries which are missing in \code{xobs} but observed in \code{xtrue}
}
