% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_categorical.R
\name{cat_to_integers}
\alias{cat_to_integers}
\title{Encode category names to integers}
\usage{
cat_to_integers(x)
}
\arguments{
\item{x}{samples of  a categorical variable}
}
\value{
A list containing
\describe{
  \item{\code{x}}{Integer encoded \code{x}}
  \item{\code{xlevels}}{Original category names corresponding to 1,2,...,k.}
}
}
\description{
For a category variable with k categories, rename the category names to 1,2,...,k in the order of appearance.
}
\keyword{internal}
