% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_fun.R
\name{comp_error}
\alias{comp_error}
\alias{cal_mae}
\alias{cal_rmse}
\alias{cal_smae}
\alias{cal_misclass}
\title{Compute imputation error}
\usage{
cal_mae(xhat, xobs, xtrue, round = FALSE)

cal_rmse(xhat, xobs, xtrue, relative = TRUE)

cal_smae(xhat, xobs, xtrue, round = FALSE, reduce = TRUE, verbose = TRUE)

cal_misclass(xhat, xobs, xtrue)
}
\arguments{
\item{xhat}{imputed data matrix or vector}

\item{xobs}{incomplete observed data matrix or vector}

\item{xtrue}{complete true data matrix or vector}

\item{round}{whether to round the values of \code{xhat} to observed integers in \code{xobs}}

\item{relative}{Only used for \code{cal_rmse}. If \code{TRUE}, normalize the computed error by the Frobenius norm of \code{xtrue}}

\item{reduce}{Only used for \code{cal_smae}. Return the average error if \code{TRUE} else the whole error vector}

\item{verbose}{Only used for \code{cal_smae}. If \code{TRUE}, throw out a warning when perfect baseline imputation appears}
}
\value{
Computed imputation error
}
\description{
Compute the imputation error on entries which are missing in \code{xobs} but observed in \code{xtrue}
}
\section{Functions}{
\itemize{
\item \code{cal_mae}: Compute the mean absolute error

\item \code{cal_rmse}: Compute the root mean squared error

\item \code{cal_smae}: Compute the MAE scaled by median imputation MAE (at each column)

\item \code{cal_misclass}: Compute the mis-classification rate for categorical data
}}

