% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fun.R
\name{continuous2ordinal}
\alias{continuous2ordinal}
\title{Transform a continuous vector to ordinal vector through cutoff function}
\usage{
continuous2ordinal(x, k = 2, cutoff = NULL)
}
\arguments{
\item{x}{continuous vector}

\item{k}{number of ordinal levels. Generated ordinal vector takes value from \code{0,\ldots,k-1}}

\item{cutoff}{cutoffs. Setting it as \code{NULL} will generate random cutoffs.}
}
\value{
an ordinal vector.
}
\description{
Discretize continuous x to ordinal z with \code{k} levels. The cutoffs are either given or randomly generated.
}
