% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{continuous2ordinal_}
\alias{continuous2ordinal_}
\title{Transform a continuous vector to ordinal vector through cutoff function}
\usage{
continuous2ordinal_(x, k = 2, by = "dist", qmin = 0.1, qmax = 0.9)
}
\arguments{
\item{x}{continuous vector}

\item{k}{number of ordinal levels. Generated ordinal vector takes value from \code{1,\ldots,k}}

\item{by}{Select cut points by absolute values if `dist` and by quantiles if `quantile`}

\item{qmin}{Cutoff points are slected in the range of `qmin` and `qmax` quantiles of the data `x`}

\item{qmax}{See `qmin`.}
}
\value{
an ordinal vector.
}
\description{
Discretize continuous x to ordinal z with \code{k} levels.
}
\keyword{internal}
