% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_impute.R
\name{ct_impute}
\alias{ct_impute}
\title{confidence intervals for Gaussian copula imputation}
\usage{
ct_impute(X, fit, alpha = 0.95)
}
\arguments{
\item{X}{The original incomplete observation}

\item{fit}{The returned object from impute_mixedgc_ppca}

\item{alpha}{The significance level}
}
\value{
A list containing two matrices \code{upper} and \code{lower} with the same size of \code{X}
\describe{
  \item{\code{upper}}{Upper bound of the constructed interval at missing location, \code{NA} as observed location.}
  \item{\code{lower}}{Lower bound of the constructed interval at missing location, \code{NA} as observed location.}
}
}
\description{
Construct confidence intervals for imputation from Gaussian copula, only applied to real-valued/continuous matrices.
}
\references{
Zhao, Y., & Udell, M. (2020). Matrix Completion with Quantified Uncertainty through Low Rank Gaussian Copula. arXiv preprint arXiv:2006.10829.
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
