% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mixedgc.R
\name{em_mixedgc}
\alias{em_mixedgc}
\title{EM algorithm to fit Gaussian copula}
\usage{
em_mixedgc(
  Z,
  Lower,
  Upper,
  d_index,
  dcat_index = NULL,
  cat_input = NULL,
  start = NULL,
  trunc_method = "Iterative",
  n_sample = 5000,
  n_update = 1,
  maxit = 50,
  eps = 0.01,
  verbose = FALSE,
  runiter = 0,
  corr_min_eigen = 0.01,
  scale_to_corr = TRUE
)
}
\arguments{
\item{Z}{Transformed latent matrix}

\item{Lower}{Lower boundary of truncated intervals}

\item{Upper}{Upper boundary of truncated intervals}

\item{d_index}{Boolean vector with \code{TRUE} at ordinal dimensions}

\item{dcat_index}{Boolean vector with \code{TRUE} at categorical dimensions}

\item{cat_input}{Input for categorical dimensions}

\item{start}{Initial value of copula correlation}

\item{trunc_method}{Method for evaluating truncated normal moments: \code{'Iterative'} or \code{'Sampling'}.}

\item{n_sample}{Number of MC samples, only used when \code{trunc_method} is \code{'Sampling'}}

\item{n_update}{The number of updates, only used when \code{trunc_method} is \code{'Iterative'}}

\item{maxit}{Maximum number of iterations}

\item{eps}{Convergence threshold}

\item{verbose}{Whether to print progress information}

\item{runiter}{When set as a positive integer, the algorithm will run the specified number of iterations exactly.}

\item{corr_min_eigen}{If the minimal eigenvalue of a correlation estimate is below \code{corr_min_eigen}, it will be regularized to have minimal eigenvalue equal to \code{corr_min_eigen}}

\item{scale_to_corr}{Whether to scale a covariance into a correlation matrix in each EM iteration. For development purpose. Use with caution.}
}
\value{
A list containing fitted copula correlation matrix, the likelihood(objective function), Z matrix with updated ordinal entries and a complete imputed Z matrix.
\describe{
  \item{\code{corr}}{Fitted copula correlation matrix}
  \item{\code{loglik}}{The log-likelihood achieved during iteration.}
  \item{\code{Z}}{Incomplete \code{Z} with approximated observed ordinal mean}
  \item{\code{Zimp}}{Complete \code{Z} with observed entries the same as \code{Zobs} and missing entries imputed}
}
}
\description{
fit the Gaussian copula model from incomplete mixed data
}
\keyword{internal}
