% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mixedgc_iter.R
\name{em_mixedgc_iter}
\alias{em_mixedgc_iter}
\title{Each iteration of EM algorithm fit}
\usage{
em_mixedgc_iter(
  Z,
  r_lower,
  r_upper,
  mu = rep(0, ncol(Z)),
  sigma,
  n_update = 1,
  trunc_method = "Iterative",
  n_sample = 5000
)
}
\arguments{
\item{Z}{Incomplete Z matrix}

\item{r_lower}{Lower boundary of truncated intervals for ordinal columns}

\item{r_upper}{Upper boundary of truncated intervals for ordinal columns}

\item{mu}{Mean vector. Default setting forces it to be \code{0}.}

\item{sigma}{Covariance matrix.}

\item{n_update}{The number of updates to conduct when `method='Iterative'`}

\item{trunc_method}{Method for evaluating truncated normal moments}

\item{n_sample}{Number of samples to use for sampling methods for evaluating truncated normal moments}
}
\value{
A list containing fitted copula correlation matrix, the likelihood(objective function), Z matrix with updated ordinal entries and a complete imputed Z matrix.
\describe{
  \item{\code{R}}{Fitted copula correlation matrix}
  \item{\code{loglik}}{The log-likelihood achieved during iteration.}
  \item{\code{Zobs}}{Incomplete \code{Z} with approximated observed ordinal entries}
  \item{\code{Zimp}}{Complete \code{Z} with observed entries the same as \code{Z} and missing entries imputed}
}
}
\description{
fit the Gaussian copula model with incomplete mixed type data
}
\references{
Zhao, Y., & Udell, M. (2020). Missing Value Imputation for Mixed Data Through Gaussian Copula. KDD 2020.
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
