% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mixedgc.R
\name{em_mixedgc_ppca}
\alias{em_mixedgc_ppca}
\title{EM algorithm to fit low rank Gaussian copula}
\usage{
em_mixedgc_ppca(
  rank,
  Z,
  Lower,
  Upper,
  d_index,
  dcat_index = NULL,
  cat_input = NULL,
  start = NULL,
  trunc_method = "Iterative",
  n_sample = 5000,
  n_update = 1,
  maxit = 50,
  eps = 0.01,
  verbose = FALSE,
  runiter = 0
)
}
\arguments{
\item{rank}{the number of latent factors}

\item{Z}{Transformed latent matrix}

\item{Lower}{Lower boundary of truncated intervals}

\item{Upper}{Upper boundary of truncated intervals}

\item{d_index}{Boolean vector with \code{TRUE} at ordinal dimensions}

\item{dcat_index}{Boolean vector with \code{TRUE} at categorical dimensions}

\item{cat_input}{Input for categorical dimensions}

\item{start}{Initial value of copula correlation}

\item{trunc_method}{Method for evaluating truncated normal moments: \code{'Iterative'} or \code{'Sampling'}.}

\item{n_sample}{Number of MC samples, only used when \code{trunc_method} is \code{'Sampling'}}

\item{n_update}{The number of updates, only used when \code{trunc_method} is \code{'Iterative'}}

\item{maxit}{Maximum number of iterations}

\item{eps}{Convergence threshold}

\item{verbose}{Whether to print progress information}

\item{runiter}{When set as a positive integer, the algorithm will run the specified number of iterations exactly.}
}
\value{
A list containing fitted copula parameters, the likelihood (objective function), Z matrix with updated ordinal entries and the conditional variance corresponding to the observed Z matrix.
\describe{
  \item{\code{W}}{Fitted latent low rank subspace matrix}
  \item{\code{sigma}}{Fitted noise variance}
  \item{\code{loglik}}{The log-likelihood achieved during iteration.}
  \item{\code{Z}}{Incomplete \code{Z} with approximated observed ordinal entries}
  \item{\code{C}}{The conditional variance corresponding to the observed Z matrix. Useful for quantifying imputation uncertainty.}
}
}
\description{
fit the Gaussian copula model from incomplete mixed data
}
\references{
Zhao, Y., & Udell, M. (2020). Matrix Completion with Quantified Uncertainty through Low Rank Gaussian Copula. arXiv preprint arXiv:2006.10829.
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
\keyword{internal}
