% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mixedgc_iter.R
\name{em_mixedgc_ppca_iter}
\alias{em_mixedgc_ppca_iter}
\title{Each iteration of EM algorithm fit (LRGC)}
\usage{
em_mixedgc_ppca_iter(Z, r_lower, r_upper, W, sigma)
}
\arguments{
\item{Z}{Incomplete Z matrix}

\item{r_lower}{Lower boundary of truncated intervals for ordinal columns}

\item{r_upper}{Upper boundary of truncated intervals for ordinal columns}

\item{W}{The latent low rank subspace matrix}

\item{sigma}{The noise variance}
}
\value{
A list containing fitted copula parameters, the likelihood (objective function), Z matrix with updated ordinal entries and the conditional variance corresponding to the observed Z matrix.
\describe{
  \item{\code{W}}{Fitted latent low rank subspace matrix}
  \item{\code{sigma}}{Fitted noise variance}
  \item{\code{loglik}}{The log-likelihood achieved during iteration.}
  \item{\code{Zobs}}{Incomplete \code{Z} with approximated observed ordinal entries}
  \item{\code{C}}{The conditional variance corresponding to the observed Z matrix. Useful for quantifying imputation uncertainty.}
  \item{\code{S}}{Required quantity to impute the Z matrix}
}
}
\description{
fit the Gaussian copula model with incomplete mixed type data
}
\references{
Zhao, Y., & Udell, M. (2020). Matrix Completion with Quantified Uncertainty through Low Rank Gaussian Copula. arXiv preprint arXiv:2006.10829.
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
