% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embody.R
\name{est_z_row_ord}
\alias{est_z_row_ord}
\title{Compute multivariate truncated normal mean and var}
\usage{
est_z_row_ord(
  z,
  lower,
  upper,
  obs_indices,
  ord_obs_indices,
  obs_in_ord,
  ord_in_obs,
  sigma_oo,
  method = "Explicit",
  n_sample = 5000
)
}
\arguments{
\item{z}{An incomplete row}

\item{lower}{Lower boundary of truncated intervals for ordinal columns}

\item{upper}{Upper boundary of truncated intervals for ordinal columns}

\item{obs_indices}{Boolean vector where \code{TRUE} indicates observed entries.}

\item{ord_obs_indices}{Boolean vector where \code{TRUE} indicates observed ordinal entries.}

\item{obs_in_ord}{Boolean vector where \code{TRUE} indicates observed ordinal entries.}

\item{ord_in_obs}{Boolean vector where \code{TRUE} indicates ordinal entries.}

\item{sigma_oo}{Sigma_{obs, obs}}

\item{method}{Explicit derivation `Explicit` or sampling method `TruncatedNormal`}

\item{n_sample}{Number of MC samples to use}
}
\value{
A list containing
\describe{
  \item{\code{mean}}{The updated mean at observed entries }
  \item{\code{var}}{The log-likelihood achieved during iteration.}
}
}
\description{
iteratively compute the conditional mean and var at observed ordinal entries
}
