% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_function.R
\name{est_z_row_ord_nocont}
\alias{est_z_row_ord_nocont}
\title{Compute multivariate truncated normal mean and covariance}
\usage{
est_z_row_ord_nocont(
  lower,
  upper,
  corr = NULL,
  obs_indices = NULL,
  cat_input = NULL,
  n_sample = 5000
)
}
\arguments{
\item{lower}{A row of \code{Lower}}

\item{upper}{A row of \code{Upper}}

\item{corr}{Current copula correlation estimate}

\item{obs_indices}{Boolean vector where \code{TRUE} indicates observed entries.}

\item{cat_input}{Input for categorical dimensions}

\item{n_sample}{Number of MC samples, only used when \code{trunc_method} is \code{'Sampling'}}
}
\value{
A list containing
\describe{
  \item{\code{mean}}{Mean for observed ordinal}
  \item{\code{cov}}{Cov for observed ordinal}
}
}
\description{
A wrapper for \code{est_z_row_ord()} where there is no continuous dimension
}
