% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{generate_copula}
\alias{generate_copula}
\alias{generate_gc}
\alias{generate_lrgc}
\title{Sample data from Gaussian copula distribution}
\usage{
generate_gc(
  n = 2000,
  var_types = NULL,
  cont_transform = NULL,
  cutoff_by = "quantile",
  qmin = 0.05,
  qmax = 0.95,
  num_ord = 5,
  corr = NULL,
  seed = NULL,
  data_only = TRUE
)

generate_lrgc(
  rank = 5,
  sigma = 0.1,
  n = 2000,
  var_types = NULL,
  cont_transform = NULL,
  cutoff_by = "quantile",
  qmin = 0.05,
  qmax = 0.95,
  num_ord = 5,
  W = NULL,
  seed = NULL,
  data_only = TRUE
)
}
\arguments{
\item{n}{number of samples to draw}

\item{var_types}{A list indicating the locations of continuous, ordinal and binary variables. The sum of its elements, \code{p}, is the desired number of variables}

\item{cont_transform}{A monotonic function to be applied to continuous columns}

\item{cutoff_by}{When ordinalizing, select cut points by absolute values if \code{dist} and by quantiles if \code{quantile}}

\item{qmin}{Cutoff points are selected in the range of \code{qmin} and \code{qmax} quantiles of the data}

\item{qmax}{See \code{qmin}}

\item{num_ord}{Number of ordinal levels to use}

\item{corr}{Only used for \code{generate_gc}. \code{NULL}, or a correlation matrix or a list of copula correlation matrices}

\item{seed}{random seed}

\item{data_only}{If \code{TRUE}, only return the generated data matrix}

\item{rank}{Only used for \code{generate_lrgc}}

\item{sigma}{Only used for \code{generate_lrgc}. A scalar between 0 and 1.}

\item{W}{Only used for \code{generate_lrgc}.  \code{NULL}, or a matrix of dimension \code{p} by \code{rank}}
}
\description{
Sample mixed data vector containing continuous, ordinal and binary data from a Gaussian copula model
}
\details{
For \code{generategc}, if \code{corr} is \code{NULL}, it will be allocated a random correlation matrix. For \code{generate_lrgc},  if \code{W} is \code{NULL}, it will be allocated a random matrix. \code{corr} used in LRGC is \code{Wt(W)+sigma*I}. If \code{corr} is a single correlation matrix, \code{n} samples will be generated from the Gaussian copula model with \code{corr} as the copula correlation. If \code{corr} is a list of correlation matrices, for each element of \code{corr}, \code{n} samples will be generated from the Gaussian copula model with that element as its copula correlation.
}
\section{Functions}{
\itemize{
\item \code{generate_gc}: Generate samples from a full rank Gaussian copula

\item \code{generate_lrgc}: Generate samples from a low rank Gaussian copula
}}

\examples{
generate_gc(n=500, var_types = list('cont'=1:5, 'ord'=6:10, 'bin'=11:15))
generate_gc(n=500, var_types = list('cont'=1:5, 'ord'=6:10), cont_transform=function(x)x^3)
generate_gc(n=500, var_types = list('cont'=1:5, 'ord'=6:10), num_ord=3)
generate_lrgc(rank = 5, sigma = 0.1, n=500, var_types = list('cont'=1:100), cont_transform=function(x)x^3)
generate_lrgc(rank = 5, sigma = 0.1, n=500, var_types = list('ord'=1:100), num_ord=10)
generate_lrgc(rank = 5, sigma = 0.1, n=500, var_types = list('bin'=1:100))
}
