% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{generate_mixed_from_gc}
\alias{generate_mixed_from_gc}
\title{Sample data from Gaussian copula distribution}
\usage{
generate_mixed_from_gc(
  sigma = NULL,
  n = 2000,
  seed = NULL,
  var_types = NULL,
  cont_transform = NULL,
  cutoff_by = "quantile",
  qmin = 0.05,
  qmax = 0.95,
  num_ord = 5
)
}
\arguments{
\item{sigma}{a correlation matrix or a list of copula correlation matrices}

\item{n}{number of samples to draw for each copula correlaiton matrix}

\item{seed}{random seed}

\item{var_types}{A list indicating the locations of continuous, ordinal and binary variables}

\item{cont_transform}{A monotonic function to be applied to continuous columns}

\item{cutoff_by}{When ordinalizing, select cut points by absolute values if `dist` and by quantiles if `quantile`}

\item{qmin}{Cutoff points are slected in the range of `qmin` and `qmax` quantiles of the data `x`}

\item{qmax}{See `qmin`.}

\item{num_ord}{Number of ordinal levels to use}
}
\value{
an ordinal vector.
}
\description{
Sample mixed data vector consisting of continuous, ordinal and binary marginals from a Gaussian copula model
}
