% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{generate_sigma}
\alias{generate_sigma}
\title{Sample random correlation matrix}
\usage{
generate_sigma(p)
}
\arguments{
\item{p}{Data dimension}
}
\value{
a \code{p*p} correlation matrix
}
\description{
Sample full rank random correlation matrix
}
