% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_mean_est.R
\name{get_cat_index_freq}
\alias{get_cat_index_freq}
\title{Compute categorical frequency}
\usage{
get_cat_index_freq(X_cat)
}
\arguments{
\item{X_cat}{An integer encoded categorical matrix}
}
\description{
For a  categorical matrix, compute the empirical frequency for each variable
}
