% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_mean_est.R
\name{get_cat_mu}
\alias{get_cat_mu}
\title{Categorical marginal estimation}
\usage{
get_cat_mu(
  freq_list,
  beta = 1000,
  n_MC = 5000,
  seed = 101,
  eps = 1e-04,
  verbose = FALSE,
  old = FALSE
)
}
\arguments{
\item{freq_list}{A list of category frequency}

\item{beta}{Constant in softmax computation}

\item{n_MC}{Number of samples used to approximate the expectation of softmax}

\item{seed}{random seed}

\item{eps}{The largest accepted marginal estimation error}

\item{verbose}{Whether to print progress information}

\item{old}{Use previous formulation?}
}
\description{
Fit the mean of Gaussian-Max distribution for given frequency
}
