% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embody.R
\name{get_trunc_2dmoments}
\alias{get_trunc_2dmoments}
\title{First two moments of truncated multivariate normal}
\usage{
get_trunc_2dmoments(
  mean,
  cov,
  lower,
  upper,
  method = "Explicit",
  n_sample = 5000
)
}
\arguments{
\item{mean}{Normal mean vector}

\item{cov}{Normal covariance matrix}

\item{lower}{Lower boundary of truncated intervals}

\item{upper}{Upper boundary of truncated intervals}

\item{method}{Method to use}

\item{n_sample}{Number of samples to use for sampling methods}
}
\value{
A list containing
\describe{
  \item{\code{mean}}{Mean vector}
  \item{\code{cov}}{Covariance matrix}
}
}
\description{
compute multivariate truncated mean vector and covariance matrix
}
