% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_fun.R
\name{grassman_dist}
\alias{grassman_dist}
\title{Compute Grassman distance between two space}
\usage{
grassman_dist(A, B)
}
\arguments{
\item{A}{A numerical matrix}

\item{B}{Another matrix that has the same dimension as \code{A}}
}
\description{
Compute Grassman distance between two space
}
\keyword{internal}
