% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mixedgc.R
\name{impute_GC}
\alias{impute_GC}
\title{Gaussian copula for incomplete mixed data}
\usage{
impute_GC(
  X,
  nlevels = 20,
  trunc_method = "Iterative",
  n_sample = 5000,
  n_update = 1,
  maxit = 50,
  eps = 0.01,
  verbose = FALSE,
  runiter = 0,
  n_MI = 0,
  corr = NULL,
  ...
)
}
\arguments{
\item{X}{A matrix or data.frame with missing values. Observed entry of \code{X} should either be numerical value or numerical ordinal level. Make sure there is no empty row nor character level in \code{X}.}

\item{nlevels}{A column which has larger number of unique values than \code{nlevels} will be classfied as continuous, otherwise ordinal.}

\item{trunc_method}{Method for evaluating truncated normal moments: \code{'Iterative'} or \code{'Sampling'}.}

\item{n_sample}{Number of MC samples, only used when \code{trunc_method} is \code{'Sampling'}}

\item{n_update}{The number of updates, only used when \code{trunc_method} is \code{'Iterative'}}

\item{maxit}{Maximum number of iterations}

\item{eps}{Convergence threshold}

\item{verbose}{Whether to print progress information}

\item{runiter}{When set as a positive integer, the algorithm will run the specified number of iterations exactly.}

\item{n_MI}{The number of random samples to draw from the missing distribution.}

\item{corr}{If not \code{NULL}, impute missing values using \code{corr} as the copula correlation}

\item{...}{Additional arguments for development use}
}
\value{
A list containing:
\describe{
  \item{\code{Ximp}}{Imputed data matrix}
  \item{\code{corr}}{Fitted copula correlation matrix}
  \item{\code{loglik}}{The log-likelihood achieved during iteration. This value approximates the true objective function we want to maximize, which is hard to evaluate. Monotonically increasing \code{loglik} sequence indicates good fit}
}
}
\description{
Fit a Gaussian copula model from (continuous and ordinal) mixed data and impute the missing entries using the fitted model
}
\details{
Impute the missing entries of continuous and ordinal mixed data by fitting a Gaussian copula model to the data.
}
\examples{
# Simulate Data
library(MASS)
# Generate 15-dim mixed data and mask 10\% observation
var_types = list('cont'=1:5, 'ord'=6:10, 'bin'=11:15)
X = generate_mixed_from_gc(var_types = var_types, n = 500)
Xmask = mask_MCAR(X, mask_fraction = 0.2)
# Fit Gaussian copula
fit = impute_mixedgc(Xmask, verbose = TRUE)

# Compute imputation Error
cal_mae_scaled(xhat = fit$Ximp, xobs = Xmask, xtrue = X)

}
\references{
Zhao, Y., & Udell, M. (2020). Missing Value Imputation for Mixed Data via Gaussian Copula. KDD 2020
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
