% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_core.R
\name{impute_init}
\alias{impute_init}
\title{Return a complete Z matrix based on SVD initialization}
\usage{
impute_init(Z_meanimp, rank, r_lower, r_upper)
}
\arguments{
\item{Z_meanimp}{Initial complete Z with mean (i.e. zero) imputation}

\item{rank}{the number of latent factors}

\item{r_lower}{Lower boundary of truncated intervals for ordinal columns}

\item{r_upper}{Upper boundary of truncated intervals for ordinal columns}
}
\value{
A complte Z matrix
}
\description{
The SVD initialization is based on a complete Z matrix with missing entries filled in by zero. Useful for getting initial estimates of copula parameters
}
\keyword{internal}
