% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mixedgc.R
\name{impute_LRGC}
\alias{impute_LRGC}
\title{Warapper}
\usage{
impute_LRGC(X, rank, maxit = 50, eps = 0.01, nlevels = 20, verbose = FALSE)

impute_LRGC(X, rank, maxit = 50, eps = 0.01, nlevels = 20, verbose = FALSE)
}
\arguments{
\item{X}{A matrix or data.frame with missing values. Observed entry of \code{X} should either be numerical value or numerical ordinal level. Make sure there is no empty row nor character level in \code{X}.}

\item{rank}{The rank, i.e. number of latent factors}

\item{maxit}{Maximum number of iterations}

\item{eps}{Convergence threshold}

\item{nlevels}{A column which has larger number of unique values than \code{nlevels} will be classfied as continuous, otherwise ordinal.}

\item{verbose}{Whether to print progress information}
}
\value{
A list containing:
\describe{
  \item{\code{Ximp}}{Imputed data matrix}
  \item{\code{W}}{Fitted latent low rank subspace matrix}
  \item{\code{sigma}}{Fitted noise variance}
  \item{\code{loglik}}{The log-likelihood achieved during iteration. This value approximates the true objective function we want to maximize, which is hard to evaluate. Monotonically increasing \code{loglik} sequence indicates good fit}
  \item{\code{Zimp}}{The imputed Z matrix. On observed ordinal entries, the entry is the corresponding estimated conditional mean. Useful for constructing confidence intervals.}
  \item{\code{C}}{The conditional variance corresponding to the observed Z matrix. Useful for quantifying imputation uncertainty.}
  \item{\code{cutoffs}}{The estimated cutoffs for ordinal dimensions. Useful for quantifying imputation uncertainty.}
}
}
\description{
Fit a low rank Gaussian copula model from (continuous and ordinal) mixed data and impute the missing entries using the fitted model
}
\details{
Impute the missing entries of continuous and ordinal mixed data by fitting a low rank Gaussian copula (LRGC) model to the data. LRGC is a subclass of Gaussian copula: it requires the copula correlation matrix to have a low rank plus diagonal decomposition: \eqn{\Sigma = WW^\top + \sigma^2 \mathrm{I}_p} where \eqn{W\in \mathbb{R}\times {p\times k}} and \eqn{k<p}.
}
\references{
Zhao, Y., & Udell, M. (2020). Matrix Completion with Quantified Uncertainty through Low Rank Gaussian Copula. NeurIPS 2020.
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
\keyword{internal}
