% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mixedgc.R
\name{impute_mixedgc}
\alias{impute_mixedgc}
\title{Impute mixed type data with Gaussian copula}
\usage{
impute_mixedgc(
  X,
  maxit = 50,
  eps = 0.01,
  nlevels = 20,
  runiter = 0,
  trunc_method = "Iterative",
  n_sample = 5000
)
}
\arguments{
\item{X}{A matrix or data.frame with missing values. Observed entry of \code{X} should either be numerical value or numerical ordinal level. Make sure there is no empty row nor character level in \code{X}.}

\item{maxit}{Maximum number of iterations}

\item{eps}{Convergence threshold}

\item{nlevels}{A column which has larger number of unique values than \code{nlevels} will be classfied as continuous, otherwise ordinal.}

\item{runiter}{When set as a positive integer, the algorithm will run \code{runiter} iterations exactly.}

\item{trunc_method}{Method for evaluating truncated normal moments}

\item{n_sample}{Number of samples to use for sampling methods for evaluating truncated normal moments}
}
\value{
A list containing:
\describe{
  \item{\code{Ximp}}{Imputed data matrix}
  \item{\code{R}}{Fitted copula correlation matrix}
  \item{\code{loglik}}{The log-likelihood achieved during iteration. This value approximates the true objective function we want to maximize, which is hard to evaluate. Monotonically increasing \code{loglik} sequence indicates good fit}
}
}
\description{
Impute the missing values of a continuous and ordinal mixed dataset through Gaussian copula model.
}
\details{
Impute the missing entries of a continuous and ordinal mixed data by fitting a Gaussian copula model to the data. The algorithm first scales original observation \code{X} to copula observation \code{Z} whose marginals are all standard normal. For continuous columns of \code{Z}, each observed entry records a value. While for ordinal columns of \code{Z}, each observed entry records an interval. The second step is to estimate the Gaussian copula correlation matrix using observed information of \code{Z}. With estimated correlation matrix, the third step imputes missing entries of \code{Z}. The last step imputes the missing entries of \code{X} based on the imputed entries of \code{Z}.
}
\examples{
# Simulate Data
library(MASS)
Sigma = matrix(c(1,.8,.8,.8,1,.64,.8,.64,1), nrow = 3)
Z = mvrnorm(n=2000, mu = rep(0,3), Sigma = Sigma)
X = Z # X has the same first column as Z, as standard normal marginal
X[,2] = continuous2ordinal(Z[,2], k = 2) # binary marginal
X[,3] = continuous2ordinal(Z[,3], k = 5) # ordinal marginal with 5 levels

# Mask 10\% Data in each column
Xobs = X
Xobs[1:100,1] = NA
Xobs[101:200,2] = NA
Xobs[201:300,3] = NA

# Fit Gaussian copula
fit = impute_mixedgc(Xobs)

# Compute imputation Error
cal_mae_scaled(xhat = fit$Ximp, xobs = Xobs, xtrue = X)

}
\references{
Zhao, Y., & Udell, M. (2019). Missing Value Imputation for Mixed Data Through Gaussian Copula. arXiv preprint arXiv:1910.12845.
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
