% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mixedgc.R
\name{impute_mixedgc_ppca}
\alias{impute_mixedgc_ppca}
\title{Impute mixed type data with low rank Gaussian copula}
\usage{
impute_mixedgc_ppca(
  X,
  rank,
  maxit = 50,
  eps = 1e-06,
  nlevels = 20,
  verbose = FALSE,
  early.stop = TRUE
)
}
\arguments{
\item{X}{A matrix or data.frame with missing values. Observed entry of \code{X} should either be numerical value or numerical ordinal level. Make sure there is no empty row nor character level in \code{X}.}

\item{rank}{the number of latent factors}

\item{maxit}{Maximum number of iterations}

\item{eps}{Convergence threshold}

\item{nlevels}{A column which has larger number of unique values than \code{nlevels} will be classfied as continuous, otherwise ordinal.}

\item{verbose}{If true, output each iteration's detail}

\item{early.stop}{If true, stop the algorithm when the likelihood change is small.}
}
\value{
A list containing:
\describe{
  \item{\code{Ximp}}{Imputed data matrix}
  \item{\code{W}}{Fitted latent low rank subspace matrix}
  \item{\code{sigma}}{Fitted noise variance}
  \item{\code{loglik}}{The log-likelihood achieved during iteration. This value approximates the true objective function we want to maximize, which is hard to evaluate. Monotonically increasing \code{loglik} sequence indicates good fit}
  \item{\code{Zimp}}{The imputed Z matrix. On observed ordinal entries, the entry is the corresponding estimated conditional mean. Useful for constructing confidence intervals.}
  \item{\code{C}}{The conditional variance corresponding to the observed Z matrix. Useful for quantifying imputation uncertainty.}
  \item{\code{cutoffs}}{The estimated cutoffs for ordinal dimensions. Useful for quantifying imputation uncertainty.}
}
}
\description{
Impute the missing values of a continuous and ordinal mixed dataset through low rank Gaussian copula model.
}
\details{
Impute the missing entries of a continuous and ordinal mixed data by fitting a low rank Gaussian copula model to the data. The algorithm first scales original observation \code{X} to copula observation \code{Z} whose marginals are all standard normal. For continuous columns of \code{Z}, each observed entry records a value. While for ordinal columns of \code{Z}, each observed entry records an interval. The second step is to estimate the Gaussian copula correlation matrix using observed information of \code{Z}. With estimated correlation matrix, the third step imputes missing entries of \code{Z}. The last step imputes the missing entries of \code{X} based on the imputed entries of \code{Z}.
}
\references{
Zhao, Y., & Udell, M. (2020). Matrix Completion with Quantified Uncertainty through Low Rank Gaussian Copula. arXiv preprint arXiv:2006.10829.
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
