% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_core.R
\name{imputeZ_mixedgc_ppca}
\alias{imputeZ_mixedgc_ppca}
\title{Impute missing entries for the Z matrix}
\usage{
imputeZ_mixedgc_ppca(Z, W, S)
}
\arguments{
\item{Z}{The returned \code{Zobs} matrix from EM algorithm}

\item{W}{The returned estimate for \code{W} matrix}

\item{S}{Returned from EM algorithm}
}
\value{
The imputed complete \code{Z} matrix
}
\description{
Using the returned quantities from EM algorithm, impute missing entries for the Z matrix
}
