% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initZ_ordinal.R
\name{initZ_ordinal}
\alias{initZ_ordinal}
\title{Simulate random value for observed ordinal entires of \code{Z}}
\usage{
initZ_ordinal(r_lower, r_upper, seed = 1)
}
\arguments{
\item{r_lower}{Lower boundary of truncated intervals for ordinal columns}

\item{r_upper}{Upper boundary of truncated intervals for ordinal columns}

\item{seed}{random seed used. Default is 1}
}
\value{
A matrix containing ordinal columns of \code{Z} whose entries are randomly generated but lie in the truncated interval.
}
\description{
Each observed ordinal entry of \code{Z} follows truncated normal distribution with mean 0, variance 1 and truncated interval provided in \code{r_lower} and \code{r_upper}. Exact values of observed ordinal entries of \code{Z} are needed to implement the EM algorith. The initial values are ramdom values from truncated normal distribution.
}
