% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration_func.R
\name{latent_operation}
\alias{latent_operation}
\title{Operation to conduct in the latent space}
\usage{
latent_operation(
  task,
  Z,
  Lower,
  Upper,
  d_index,
  dcat_index,
  cat_input,
  corr,
  trunc_method = "Iterative",
  n_update = 1,
  n_sample = 5000,
  n_MI = 1
)
}
\arguments{
\item{task}{Task to perform. One of \code{"em", "fillup", "sample"}.}

\item{Z}{Transformed latent matrix}

\item{Lower}{Lower boundary of truncated intervals}

\item{Upper}{Upper boundary of truncated intervals}

\item{d_index}{Boolean vector with \code{TRUE} at ordinal dimensions}

\item{dcat_index}{Boolean vector with \code{TRUE} at categorical dimensions}

\item{cat_input}{Input for categorical dimensions}

\item{corr}{Current copula correlation estimate}

\item{trunc_method}{Method for evaluating truncated normal moments: \code{'Iterative'} or \code{'Sampling'}.}

\item{n_update}{The number of updates, only used when \code{trunc_method} is \code{'Iterative'}}

\item{n_sample}{Number of MC samples, only used when \code{trunc_method} is \code{'Sampling'}}

\item{n_MI}{The number of random samples to draw from the missing distribution.}
}
\value{
A list containing
\describe{
  \item{\code{corr}}{Available when \code{task = 'em'}. Updated correlation estimate.}
  \item{\code{loglik}}{Available when \code{task = 'em'}. The average log-likelihood.}
  \item{\code{Z}}{Available when \code{task = 'em'}. Incomplete \code{Z} with updated observed ordinal entries}
  \item{\code{Zimp}}{Available when \code{task = 'em'} or \code{task == 'fillup'} . Complete \code{Z} with observed entries the same as \code{Z} and missing entries imputed}
  \item{\code{Zimp_sample}}{Available when \code{task = 'sample'}. Multiple imputation samples.}
  \item{\code{C}}{Available when \code{task = 'em'}. The conditional co-variance due to missingness}
  \item{\code{var_ordinal}}{Available when \code{task = 'em'} or \code{task = 'fillup'}. The conditional variance due to truncation, i.e. Var(z|a < z < b)}
}
}
\description{
Conduct one of three tasks in the latent space. If \code{task = "em"}, conduct an em iteration. If \code{task = "fillup"}, impute the missing entries of \code{Z} as their conditional mean. If \code{task = "sample"}, conduct multiple imputation on the missing entries of \code{Z}.
}
\keyword{internal}
