% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_iteration_func.R
\name{latent_operation_row}
\alias{latent_operation_row}
\title{Operation for each latent row}
\usage{
latent_operation_row(
  z,
  lower,
  upper,
  mu,
  sigma,
  n_update = 1,
  trunc_method = "Iterative",
  n_sample = 5000
)
}
\arguments{
\item{z}{Incomplete Z matrix}

\item{lower}{Lower boundary of truncated intervals for ordinal columns}

\item{upper}{Upper boundary of truncated intervals for ordinal columns}

\item{mu}{Mean vector. Default setting forces it to be \code{0}.}

\item{sigma}{Covariance matrix.}

\item{n_update}{The number of updates to conduct when `method='Iterative'`}

\item{trunc_method}{Method for evaluating truncated normal moments}

\item{n_sample}{Number of samples to use for sampling methods for evaluating truncated normal moments}
}
\value{
A list containing fitted copula correlation matrix, the likelihood(objective function), Z matrix with updated ordinal entries and a complete imputed Z matrix.
\describe{
  \item{\code{loglik}}{The average log-likelihood achieved during iteration.}
  \item{\code{Z}}{Incomplete \code{Z} with approximated observed ordinal entries}
  \item{\code{Zimp}}{Complete \code{Z} with observed entries the same as \code{Z} and missing entries imputed}
  \item{\code{C}}{The conditional co-variance due to missingness, i.e. E(z_i z_j|x_i x_j)}
  \item{\code{var_ordinal}}{The conditional variance due to truncation, i.e. E(z|a < z < b)}
}
}
\description{
operation for each latent row
}
