% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_mask.R
\name{mask_MCAR}
\alias{mask_MCAR}
\title{Mask data}
\usage{
mask_MCAR(X, mask_fraction, seed = NULL, silence_rate = 0.01, mask_cols = NULL)
}
\arguments{
\item{X}{data to be masked}

\item{mask_fraction}{Fraction of observed entries to be masked}

\item{seed}{Seed for mask}

\item{silence_rate}{The random masking is done while ensuring there is no empty row, which may lead to smaller than specified \code{mask_fraction}. If the difference is more than \code{silence_rate}, a message will be printed}

\item{mask_cols}{If not \code{NULL}, the masking only happens in columns \code{mask_cols}}
}
\value{
masked data
}
\description{
Masks \code{mask_fraction} entries of \code{X} completely at random
}
