% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_normal.R
\name{mean_tnorm}
\alias{mean_tnorm}
\alias{var_tnorm}
\title{Truncated normal mean and variance}
\usage{
mean_tnorm(mu, sigma, a, b)

var_tnorm(mu, sigma, a, b)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{Left endpoint of the truncated interval}

\item{b}{Right enpoint of the truncated interval}
}
\description{
Compute the mean and variance of one dimensional truncated normal random variable
}
