% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_categorical.R
\name{project_to_nominal_corr}
\alias{project_to_nominal_corr}
\title{Copula correlation projection}
\usage{
project_to_nominal_corr(sigma, cat_index_list, eps = 1e-05)
}
\arguments{
\item{sigma}{copula  correlation}

\item{cat_index_list}{list indicating the number of levels for each categorical variable}

\item{eps}{minimal allowed eigenvalue}
}
\description{
Project a copula correlation matrix to be the identity at each categorical block
}
\keyword{internal}
