% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fun.R
\name{range_transform}
\alias{range_transform}
\title{Transform original observation X to copula observation Z}
\usage{
range_transform(X, type)
}
\arguments{
\item{X}{original observation matrix}

\item{type}{Speficy data type. Either \code{continuous} or \code{ordinal}.}
}
\value{
A list containing
\describe{
  \item{\code{r_val}}{Return when \code{type=continuous}. Corresponding value}
  \item{\code{r_lower}}{Return when \code{type=ordinal}. Lower boundary of corresponding truncated interval}
  \item{\code{r_upper}}{Return when \code{type=ordinal}. Upper boundary of corresponding truncated interval}
}
}
\description{
For each continuous column of \code{X}, compute its corresonding value of \code{Z}; For each ordinal column of \code{Z}, compute its corresponding truncated interval of \code{Z}.
}
