% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_reliability.R
\name{reliability_cont}
\alias{reliability_cont}
\title{Compute reliability}
\usage{
reliability_cont(X, fit, alpha = 0.95, type = 1)
}
\arguments{
\item{X}{Incomplete data}

\item{fit}{Fitted LRGC model on data \code{X}}

\item{alpha}{The confidence leveled used in computing reliability}

\item{type}{Default \code{1} corresponds to the one used in our paper}
}
\description{
Compute reliability for continuous matrices
}
