% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_reliability.R
\name{reliability_ord}
\alias{reliability_ord}
\title{Compute reliability}
\usage{
reliability_ord(X, fit, cuts.fix = NULL)
}
\arguments{
\item{X}{Incomplete data}

\item{fit}{Fitted LRGC model on data \code{X}}

\item{cuts.fix}{If \code{NULL}, use the cut points estimated from \code{fit}}
}
\description{
Compute reliability for ordinal matrices
}
