% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxi_LRGC.R
\name{scale_corr}
\alias{scale_corr}
\title{Adjust estiamted correlation parameters}
\usage{
scale_corr(W, sigma)
}
\arguments{
\item{W}{The latent low rank subspace matrix (initial estimate)}

\item{sigma}{The noise variance (initial estimate)}
}
\value{
A list containing
\describe{
  \item{\code{W}}{Adjusted latent low rank subspace matrix}
  \item{\code{sigma}}{Adjusted noise variance}
}
}
\description{
adjust estiamted correlation parameters, W and sigma, to satisfy unit diagonal constraints.
}
