% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_mask.R
\name{split_mask_val_test}
\alias{split_mask_val_test}
\title{Split observation into training  and validation}
\usage{
split_mask_val_test(X_mask, X, val_ratio = 0.5, seed = NULL)
}
\arguments{
\item{X_mask}{Masked data}

\item{X}{Original data}

\item{val_ratio}{The ratio of validation to training entries}

\item{seed}{Seed for mask}
}
\description{
Split observation into training  and validation
}
