% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxi_LRGC.R
\name{sum_3d_scale}
\alias{sum_3d_scale}
\title{Sum a 3d matrix along one axis}
\usage{
sum_3d_scale(M, c, index)
}
\arguments{
\item{M}{A 3d matrix with dimension n*k*k}

\item{c}{A vector of length n}

\item{index}{a subset of \{1,2,...,n\}}
}
\value{
A k*k matrix
}
\description{
For a 3d matrix M with dimension n*k*k, sum over dimension (2,3) with scale c, at entries selected by index in dimension 1
}
