% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_core.R
\name{to_numeric_matrix}
\alias{to_numeric_matrix}
\title{DataFrame to Matrix}
\usage{
to_numeric_matrix(X)
}
\arguments{
\item{X}{a data frame}
}
\description{
Safelly turn a data frame to a numerical matrix
}
