% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fun.R
\name{Ximp_transform}
\alias{Ximp_transform}
\title{Transform imputed Z matrix to original scale X}
\usage{
Ximp_transform(Z, X, r_upper, d_index)
}
\arguments{
\item{Z}{imputed Z matrix}

\item{X}{original observation matrix}

\item{r_upper}{Speficy data type. Either \code{continuous} or \code{ordinal}.}

\item{d_index}{ordinal dimension indexes.}
}
\value{
Imputed data matrix
}
\description{
Impute orginal incomplete observation using imputed \code{Z} matrix and marginal distribution information drawn from original observation \code{X}
}
