% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_categorical.R
\name{Z_to_original_trunc}
\alias{Z_to_original_trunc}
\title{Transform latent for cat}
\usage{
Z_to_original_trunc(Z, X_cat, cat_index_list, old = FALSE)
}
\arguments{
\item{Z}{Matrix corresponding to ONLY latent categorical variables}

\item{X_cat}{(incomplete) categorical data matrix}

\item{cat_index_list}{list indicating the number of levels for each categorical variable}

\item{old}{Use previous stringent categorical model?}
}
\description{
Transform latent matrix corresponding to latent categorical variables so that after transformation, interval truncated gaussian is achieved
}
