% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_fun.R
\name{cal_mae}
\alias{cal_mae}
\title{compute the mean absolute error}
\usage{
cal_mae(xhat, xobs = NULL, xtrue, round = FALSE)
}
\arguments{
\item{xhat}{imputed data matrix or vector}

\item{xobs}{incomplete observed data matrix or vector}

\item{xtrue}{complete true data matrix or vector}

\item{round}{whether round the values of \code{xhat} to observed integers}
}
\value{
computed error
}
\description{
compute the mean absolute error on entries which are missing in \code{xobs} but observed in \code{xtrue}
}
