% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_fun.R
\name{cal_rmse}
\alias{cal_rmse}
\title{compute root mean squared error}
\usage{
cal_rmse(xhat, xtrue)
}
\arguments{
\item{xhat}{imputed value vector}

\item{xtrue}{true value vector}
}
\value{
computed error
}
\description{
compute root mean squared error
}
