% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{continuous2ordinal}
\alias{continuous2ordinal}
\title{Transform a continuous vector to ordinal vector through cutoff function.}
\usage{
continuous2ordinal(x, k = 2, by = "dist", qmin = 0.1, qmax = 0.9, max_try = 20)
}
\arguments{
\item{x}{continuous vector}

\item{k}{number of ordinal levels. Generated ordinal vector takes value from \code{1,\ldots,k}}

\item{by}{Select cut points by absolute values if `dist` and by quantiles if `quantile`}

\item{qmin}{Cutoff points are slected in the range of `qmin` and `qmax` quantiles of the data `x`}

\item{qmax}{See `qmin`.}

\item{max_try}{The maximum number of attempts to select random cutoffs, since some may yield to ordinal variable with fewer than \code{k} levels}
}
\value{
an ordinal vector.
}
\description{
Discretize continuous x to ordinal z with \code{k} levels. If the generated values are guaranteed to have \code{k} unique levels, otherwise, error will be raised.
}
