% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mixedgc.R
\name{em_mixedgc}
\alias{em_mixedgc}
\title{EM algorithm to fit Gaussian copula}
\usage{
em_mixedgc(Z_continuous, r_lower, r_upper, start = NULL, maxit = 100,
  eps = 0.001, stop.relative = FTRUE)
}
\arguments{
\item{Z_continuous}{Continuous columns}

\item{r_lower}{Lower boundary of truncated intervals for ordinal columns}

\item{r_upper}{Upper boundary of truncated intervals for ordinal columns}

\item{start}{Initial value of copula correlation matrix. Default is \code{NULL}.}

\item{maxit}{maximum number of iterations}

\item{eps}{Convergence threshold}

\item{stop.relative}{Use relative Frobeinus error when \code{TRUE} and maximum elementise error when \code{FALSE} by default}
}
\value{
A list containing fitted copula correlation matrix, the likelihood(objective function), Z matrix with updated ordinal entries and a complete imputed Z matrix.
\describe{
  \item{\code{R}}{Fitted copula correlation matrix}
  \item{\code{loglik}}{The log-likelihood achieved during iteration.}
  \item{\code{Zobs}}{Incomplete \code{Z} with approximated observed ordinal entries}
  \item{\code{Zimp}}{Complete \code{Z} with observed entries the same as \code{Zobs} and missing entries imputed}
}
}
\description{
fit the Gaussian copula model from incomplete mixed data
}
\references{
Zhao, Y., & Udell, M. (2019). Missing Value Imputation for Mixed Data Through Gaussian Copula. arXiv preprint arXiv:1910.12845.
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
