% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mixedgc.R
\name{em_mixedgc}
\alias{em_mixedgc}
\title{EM algorithm to fit Gaussian copula}
\usage{
em_mixedgc(
  Z_continuous,
  r_lower,
  r_upper,
  start = NULL,
  maxit = 100,
  eps = 0.001,
  runiter = 0,
  trunc_method = "Iterative",
  n_sample = 5000,
  n_update = 1,
  verbose = FALSE,
  scale_to_corr = TRUE
)
}
\arguments{
\item{Z_continuous}{Continuous columns}

\item{r_lower}{Lower boundary of truncated intervals for ordinal columns}

\item{r_upper}{Upper boundary of truncated intervals for ordinal columns}

\item{start}{Initial value of copula correlation matrix. Default is \code{NULL}.}

\item{maxit}{maximum number of iterations}

\item{eps}{Convergence threshold}

\item{runiter}{When set as a positive integer, the algorithm will run \code{runiter} iterations exactly.}

\item{trunc_method}{Method for evaluating truncated normal moments}

\item{n_sample}{Number of samples to use for sampling methods for evaluating truncated normal moment. Only used when `trunc_method='Sampling_TN'`}

\item{n_update}{Number of iterative updates to conduct. Only used when `trunc_method='Iterative'`}

\item{verbose}{Whether to print progress information}

\item{scale_to_corr}{Whether to scale a covariance into a correlation matrix in each EM iteration. For development purpose. Use with caution.}
}
\value{
A list containing fitted copula correlation matrix, the likelihood(objective function), Z matrix with updated ordinal entries and a complete imputed Z matrix.
\describe{
  \item{\code{corr}}{Fitted copula correlation matrix}
  \item{\code{loglik}}{The log-likelihood achieved during iteration.}
  \item{\code{Z}}{Incomplete \code{Z} with approximated observed ordinal entries}
  \item{\code{Zimp}}{Complete \code{Z} with observed entries the same as \code{Zobs} and missing entries imputed}
}
}
\description{
fit the Gaussian copula model from incomplete mixed data
}
