% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mixedgc.R
\name{em_mixedgc_ppca}
\alias{em_mixedgc_ppca}
\title{EM algorithm to fit low rank Gaussian copula}
\usage{
em_mixedgc_ppca(
  rank,
  Z_continuous,
  r_lower,
  r_upper,
  start = NULL,
  maxit = 100,
  eps = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{rank}{the number of latent factors}

\item{Z_continuous}{Continuous columns}

\item{r_lower}{Lower boundary of truncated intervals for ordinal columns}

\item{r_upper}{Upper boundary of truncated intervals for ordinal columns}

\item{start}{Initial value of copula correlation matrix. Default is \code{NULL}.}

\item{maxit}{maximum number of iterations}

\item{eps}{Convergence threshold}

\item{verbose}{If true, output each iteration's detail}
}
\value{
A list containing fitted copula parameters, the likelihood (objective function), Z matrix with updated ordinal entries and the conditional variance corresponding to the observed Z matrix.
\describe{
  \item{\code{W}}{Fitted latent low rank subspace matrix}
  \item{\code{sigma}}{Fitted noise variance}
  \item{\code{loglik}}{The log-likelihood achieved during iteration.}
  \item{\code{Zobs}}{Incomplete \code{Z} with approximated observed ordinal entries}
  \item{\code{C}}{The conditional variance corresponding to the observed Z matrix. Useful for quantifying imputation uncertainty.}
  \item{\code{S}}{Required quantity to impute the Z matrix}
}
}
\description{
fit the Gaussian copula model from incomplete mixed data
}
\references{
Zhao, Y., & Udell, M. (2020). Matrix Completion with Quantified Uncertainty through Low Rank Gaussian Copula. arXiv preprint arXiv:2006.10829.
}
\author{
Yuxuan Zhao, \email{yz2295@cornell.edu} and Madeleine Udell, \email{udell@cornell.edu}
}
\keyword{internal}
