% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_function.R
\name{est_z_row_ord}
\alias{est_z_row_ord}
\title{Compute the observed ordinal mean and cov in a row}
\usage{
est_z_row_ord(
  z,
  lower,
  upper,
  sigma_oo,
  ord_indices = NULL,
  obs_indices = NULL,
  ord_obs_indices = NULL,
  obs_in_ord = NULL,
  ord_in_obs = NULL,
  n_sample = 5000
)
}
\arguments{
\item{z}{A row of \code{Z}}

\item{lower}{A row of \code{Lower}}

\item{upper}{A row of \code{Upper}}

\item{sigma_oo}{\eqn{Sigma_{obs, obs}}}

\item{ord_indices}{Boolean vector where \code{TRUE} indicates ordinal entries.}

\item{obs_indices}{Boolean vector where \code{TRUE} indicates observed entries.}

\item{ord_obs_indices}{Boolean vector where \code{TRUE} indicates observed ordinal entries.}

\item{obs_in_ord}{Boolean vector where \code{TRUE} indicates observed ordinal entries.}

\item{ord_in_obs}{Boolean vector where \code{TRUE} indicates ordinal entries.}

\item{n_sample}{Number of MC samples, only used when \code{trunc_method} is \code{'Sampling'}}
}
\value{
A list containing
\describe{
  \item{\code{mean}}{Mean for observed ordinal}
  \item{\code{cov}}{Cov for observed ordinal}
}
}
\description{
Estimate the conditional mean and cov at observed ordinal entries through sampling
}
\keyword{internal}
