% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_core.R
\name{get_cond_dist}
\alias{get_cond_dist}
\title{Conditional normal mean and cov}
\usage{
get_cond_dist(
  z,
  mu,
  cov,
  index_o,
  index_m = NULL,
  test_logical = TRUE,
  drop = TRUE
)
}
\arguments{
\item{z}{data observation}

\item{mu}{normal mean}

\item{cov}{normal covariance}

\item{index_o}{observed dimensions}

\item{index_m}{missing dimensions}
}
\value{
A list containing
\describe{
  \item{\code{mean}}{conditional mean}
  \item{\code{cov}}{conditional covariance }
}
}
\description{
Compute the conditional normal mean and cov
}
