% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initZ_ordinal.R
\name{initZ_ordinal}
\alias{initZ_ordinal}
\title{Generate values for latent \code{Z} at observed ordinal entires}
\usage{
initZ_ordinal(Lower, Upper, seed = NULL, method = "univariate_mean")
}
\arguments{
\item{Lower}{Lower boundary of truncated intervals}

\item{Upper}{Upper boundary of truncated intervals}

\item{seed}{random seed used}

\item{method}{method for initializing the mean}
}
\value{
Matrix \code{Z} with the same shape as \code{lower}
}
\description{
Each observed ordinal entry of \code{Z} follows truncated normal distribution with mean 0, variance 1 and truncated interval provided in \code{lower} and \code{upper}. Fill out those entries by their univariate mean if \code{method = "univariate_mean"} and by random sampples if \code{method = "sampling"}
}
