% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_mask.R
\name{mask_MCAR}
\alias{mask_MCAR}
\title{Mask data}
\usage{
mask_MCAR(X, mask_fraction, max_try = 50, seed = NULL, allow_empty_row = FALSE)
}
\arguments{
\item{X}{data to be masked}

\item{mask_fraction}{Fraction of observed entries to be masked}

\item{max_try}{Number of attempts to conduct mask, if an empty row appears after a masking attempt}

\item{seed}{Seed for mask}

\item{allow_empty_row}{If False, allow some masked rows to be empty}
}
\value{
masked data
}
\description{
Masks \code{mask_fraction} entries of \code{X} completely at random
}
