% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_function.R
\name{update_z_row_ord}
\alias{update_z_row_ord}
\title{Compute multivariate truncated normal mean and var}
\usage{
update_z_row_ord(
  z,
  lower,
  upper,
  obs_indices,
  ord_obs_indices,
  ord_in_obs,
  obs_in_ord,
  f_sigma_oo_inv_z,
  sigma_oo_inv_diag,
  n_update = 1
)
}
\arguments{
\item{z}{A row of \code{Z}}

\item{lower}{A row of \code{Lower}}

\item{upper}{A row of \code{Upper}}

\item{obs_indices}{Boolean vector where \code{TRUE} indicates observed entries.}

\item{ord_obs_indices}{Boolean vector where \code{TRUE} indicates observed ordinal entries.}

\item{ord_in_obs}{Boolean vector where \code{TRUE} indicates ordinal entries.}

\item{obs_in_ord}{Boolean vector where \code{TRUE} indicates observed ordinal entries.}

\item{f_sigma_oo_inv_z}{A function computing The matrix-vector product \eqn{Sigma_{obs, obs}^{-1} * z_{obs}}}

\item{sigma_oo_inv_diag}{The diagonal of \eqn{Sigma_{obs, obs}^{-1}}}

\item{n_update}{The number of updates, only used when \code{trunc_method} is \code{'Iterative'}}
}
\value{
A list containing
\describe{
  \item{\code{mean}}{Mean for observed ordinal}
  \item{\code{var}}{Var for observed ordinal}
}
}
\description{
iteratively compute the conditional mean and var at observed ordinal entries
}
