% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fun.R
\name{Ximp_transform}
\alias{Ximp_transform}
\title{Transform imputed Z matrix to X in the observed space}
\usage{
Ximp_transform(Z, X, d_index, c_index = NULL)
}
\arguments{
\item{Z}{imputed Z matrix}

\item{X}{original observation matrix}

\item{d_index}{ordinal dimension indexes.}
}
\value{
Imputed data matrix
}
\description{
Impute orginal incomplete observation using imputed \code{Z} matrix and marginal distribution information drawn from original observation \code{X}
}
